secrets masking functions cleanup
